% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_bingads_data.R
\docType{data}
\name{my_bingads_data}
\alias{my_bingads_data}
\title{Sample of digital marketing data from bing Ads downloaded by means of the Windsor.ai API.}
\format{
A data frame with 164 rows and 5 variables:
\describe{
  \item{campaign}{name of the campaign}
  \item{clicks}{number of clicks}
  \item{spend}{spend data}
  \item{impressions}{impressions data}
  \item{date}{date}
}
}
\source{
\url{https://windsor.ai/}
}
\usage{
my_bingads_data
}
\description{
A dataset containing sample bing Ads data
fetched from windsor.ai API. See more at:  https://windsor.ai/
}
\keyword{datasets}
